rename generic to generic_tech
