+++ gdsfactory/technology/generic_tech/klayout/icon_128x128.png
