+++ gdsfactory/technology/generic_tech/klayout/pymacros/set_shortcuts.lym
